Fix NULL ptr dereferences
NULL pointer dereferencing is -- unfortunately -- undefined behaviour in
C and is not guaranteed to generate a SEGFAULT.

This commit fixes a couple of possible NULL pointer dereferences across
the whole codebase, which I caught using static analysis.

+++ daemon/c_vol.c
@@ -371,11 +371,14 @@ c_vol_btrfs_create_subvol(const char *dev, const char *mount_data)
 		ERROR_ERRNO("Could not umount temporary mount of btrfs root volume %s!", dev);
 	}
 out:
-	unlink(tmp_mount);
+	if (tmp_mount)
+		unlink(tmp_mount);
 	if (subvol_path)
 		mem_free(subvol_path);
-	mem_free(tmp_mount);
-	mem_free(token);
+	if (tmp_mount)
+		mem_free(tmp_mount);
+	if (token)
+		mem_free(token);
 	return ret;
 }
